BETTER REVIEW - This makes major change to the DB schema
I think we made a mistake choosing TEXT fields to be primary keys, so I changed anything that was a TEXT value to an INT for a primary key. For instance Medication and Illness tables now have a new column for primary key.

I changed the primary key in New_Applicant_Form to only be P_SSN, instead of Email and P_SSN. This just seems redundant.

This changes the schema so has the potential to affect components you guys may have been working on if they are querying the database for tables that will change with this update!